Looking down for post and comment mutation

+++ src/resolvers/Mutation.js
@@ -36,12 +36,13 @@ const Mutation = {
 
     return prisma.mutation.createPost({ data }, info);
   },
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     const data = {
       ...args.data,
       author: {
         connect: {
-          id: args.data.author
+          id: userId
         }
       },
       post: {
@@ -80,7 +81,19 @@ const Mutation = {
       info
     );
   },
-  async deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error('Unable delete this comment');
+
+
     return prisma.mutation.deleteComment({ where: { id: args.id } }, info);
   },
   async updateUser(parent, args, { prisma, request }, info) {
@@ -88,10 +101,32 @@ const Mutation = {
 
     return prisma.mutation.updateUser({ where: { id: userId }, data: args.data }, info)
   },
-  async updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) throw new Error('Unable update this post');
+
     return prisma.mutation.updatePost({ where: { id: args.id }, data: args.data }, info);
   },
-  async updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if (!commentExists) throw new Error('Unable update this comment');
+
     return prisma.mutation.updateComment({ where: { id: args.id }, data: args.data }, info);
   },
   async login(parent, args, { prisma }, info) {
